Added restcontroller

+++ src/main/java/ru/otus/spring/service/BookServiceImpl.java
@@ -2,14 +2,16 @@ package ru.otus.spring.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.otus.spring.domain.dto.AuthorDto;
+import ru.otus.spring.domain.dto.BookDto;
+import ru.otus.spring.domain.dto.GenreDto;
 import ru.otus.spring.repository.BookRepository;
-import ru.otus.spring.domain.Author;
 import ru.otus.spring.domain.Book;
-import ru.otus.spring.domain.Genre;
 import ru.otus.spring.exception.BookNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class BookServiceImpl implements BookService {
@@ -17,52 +19,58 @@ public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
     private final AuthorService authorService;
     private final GenreService genreService;
+    private final MappingService mappingService;
 
-    public BookServiceImpl(BookRepository bookRepository, AuthorService authorService, GenreService genreService) {
+    public BookServiceImpl(BookRepository bookRepository, AuthorService authorService, GenreService genreService, MappingService mappingService) {
         this.bookRepository = bookRepository;
         this.authorService = authorService;
         this.genreService = genreService;
+        this.mappingService = mappingService;
     }
 
     @Transactional
     @Override
-    public Book insert(Book book) {
-        if (book.getId() != null && book.getId() != 0)
+    public BookDto insert(BookDto bookDto) {
+        if (bookDto.getId() != null && bookDto.getId() != 0)
             throw new RuntimeException("При добавлении книги идентификатор должен быть пустым");
-        return save(book);
+        return save(bookDto);
     }
 
     @Transactional
     @Override
-    public Book update(Book book) {
-        Long bookId = book.getId();
-        Long cnt = bookRepository.countById(bookId);
-        if (cnt.equals(0L)) {
+    public BookDto update(BookDto bookDto) {
+        Long bookId = bookDto.getId();
+
+        if (bookRepository.countById(bookId) == 0)
             throw new BookNotFoundException(bookId);
-        }
-        return save(book);
+
+        return save(bookDto);
     }
 
-    private Book save(Book book) {
-        Author author = Optional.ofNullable(book.getAuthor()).orElse(new Author());
-        Genre genre = Optional.ofNullable(book.getGenre()).orElse(new Genre());
+    private BookDto save(BookDto bookDto) {
+        AuthorDto authorDto = Optional.ofNullable(bookDto.getAuthor()).orElse(new AuthorDto());
+        GenreDto genreDto = Optional.ofNullable(bookDto.getGenre()).orElse(new GenreDto());
 
-        if (author.getId() == null || author.getId() == 0)
-            author = authorService.getOrSaveByName(author.getName());
+        if (authorDto.getId() == null || authorDto.getId() == 0)
+            authorDto = mappingService.toDto(authorService.getOrSaveByName(authorDto.getName()));
 
-        if (genre.getId() == null || genre.getId() == 0)
-            genre = genreService.getOrSaveByName(genre.getName());
+        if (genreDto.getId() == null || genreDto.getId() == 0)
+            genreDto = mappingService.toDto(genreService.getOrSaveByName(genreDto.getName()));
 
-        return bookRepository.save(Book.builder()
-                .id(book.getId())
-                .name(book.getName())
-                .author(author)
-                .genre(genre)
+        Book saved = bookRepository.save(Book.builder()
+                .id(bookDto.getId())
+                .name(bookDto.getName())
+                .author(mappingService.toDomainObject(authorDto))
+                .genre(mappingService.toDomainObject(genreDto))
                 .build());
+
+        return mappingService.toDto(saved);
     }
 
     @Override
     public void deleteById(Long bookId) {
+        if (bookRepository.countById(bookId) == 0)
+            throw new BookNotFoundException(bookId);
         bookRepository.deleteById(bookId);
     }
 
@@ -73,6 +81,12 @@ public class BookServiceImpl implements BookService {
             return bookOpt.get();
         else
             throw new BookNotFoundException(bookId);
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public BookDto findByIdInDto(Long bookId) {
+        return mappingService.toDto(findById(bookId));
     }
 
     @Override
@@ -87,5 +101,14 @@ public class BookServiceImpl implements BookService {
     @Override
     public List<Book> findAll() {
         return bookRepository.findAll();
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public List<BookDto> findAllInDto() {
+        return findAll()
+                .stream()
+                .map(mappingService::toDto)
+                .collect(Collectors.toList());
     }
 }
